+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix IDE freeze caused by duplicate background loading tasks and stuck loading state on cancellation
+
 ## [0.3.1] - 2026-02-24
 
 ### Fixed